# kura/explorer/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kura Explorer API",
    "description": "API for exploring Kura checkpoint data",
    "version": "1.0.0"
  },
  "paths": {
    "/api/clusters": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "Get Clusters",
        "description": "Get clusters with pagination and filtering.",
        "operationId": "get_clusters_api_clusters_get",
        "parameters": [
          {
            "name": "parent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by parent cluster ID",
              "title": "Parent Id"
            },
            "description": "Filter by parent cluster ID"
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by cluster level",
              "title": "Level"
            },
            "description": "Filter by cluster level"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(name|conversation_count|frustration)$",
              "default": "name",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_desc",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Sort Desc"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by language",
              "title": "Language"
            },
            "description": "Filter by language"
          },
          {
            "name": "min_frustration",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 5.0,
                  "minimum": 1.0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Frustration"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/clusters/{cluster_id}": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "Get Cluster",
        "description": "Get detailed cluster information.",
        "operationId": "get_cluster_api_clusters__cluster_id__get",
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/clusters/{cluster_id}/summary": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "Get Cluster Summary",
        "description": "Get aggregated summary of a cluster.",
        "operationId": "get_cluster_summary_api_clusters__cluster_id__summary_get",
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ClusterDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "x_coord": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X Coord"
          },
          "y_coord": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y Coord"
          },
          "conversation_count": {
            "type": "integer",
            "title": "Conversation Count"
          },
          "avg_frustration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Frustration"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          },
          "parent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClusterResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/ClusterResponse"
            },
            "type": "array",
            "title": "Children"
          },
          "sample_conversations": {
            "items": {
              "$ref": "#/components/schemas/ConversationResponse"
            },
            "type": "array",
            "title": "Sample Conversations"
          },
          "hierarchy": {
            "items": {
              "$ref": "#/components/schemas/ClusterResponse"
            },
            "type": "array",
            "title": "Hierarchy"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "level",
          "conversation_count"
        ],
        "title": "ClusterDetailResponse",
        "description": "Detailed cluster info."
      },
      "ClusterResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "x_coord": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X Coord"
          },
          "y_coord": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y Coord"
          },
          "conversation_count": {
            "type": "integer",
            "title": "Conversation Count",
            "default": 0
          },
          "child_count": {
            "type": "integer",
            "title": "Child Count",
            "default": 0
          },
          "avg_frustration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Frustration"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "level"
        ],
        "title": "ClusterResponse",
        "description": "Basic cluster info."
      },
...
      "ConversationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          },
          "message_count": {
            "type": "integer",
            "title": "Message Count"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SummaryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "cluster_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cluster Names"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "metadata",
          "message_count"
        ],
        "title": "ConversationResponse",
        "description": "Basic conversation info."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse": {
        "properties": {
          "items": {
            "items": {},
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "PaginatedResponse",
        "description": "Paginated response wrapper."
      },
...
      "SummaryResponse": {
        "properties": {
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "request": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request"
          },
          "languages": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Languages"
          },
          "task": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task"
          },
          "concerning_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concerning Score"
          },
          "user_frustration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Frustration"
          },
          "assistant_errors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assistant Errors"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "chat_id",
          "summary"
        ],
        "title": "SummaryResponse",
        "description": "Conversation summary."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}